+++ coverage/control.py
@@ -846,7 +846,8 @@ class Coverage(object):
     def report(
         self, morfs=None, show_missing=None, ignore_errors=None,
         file=None, omit=None, include=None, skip_covered=None,
-        contexts=None, skip_empty=None, precision=None, sort=None
+        contexts=None, skip_empty=None, precision=None, sort=None,
+        always_total=None
     ):
         """Write a textual summary report to `file`.
 
@@ -891,13 +892,16 @@ class Coverage(object):
         .. versionadded:: 5.2
             The `precision` parameter.
 
+        .. versionadded:: 5.4
+            The `always_total` parameter.
+
         """
         with override_config(
             self,
             ignore_errors=ignore_errors, report_omit=omit, report_include=include,
             show_missing=show_missing, skip_covered=skip_covered,
             report_contexts=contexts, skip_empty=skip_empty, precision=precision,
-            sort=sort
+            sort=sort, always_total=always_total
         ):
             reporter = SummaryReporter(self)
             return reporter.report(morfs, outfile=file)