+++ coverage/summary.py
@@ -120,8 +120,9 @@ class SummaryReporter(object):
         for line in lines:
             self.writeout(line[0])
 
-        # Write a TOTAl line if we had more than one file.
-        if self.total.n_files > 1:
+        # Write a TOTAl line if we had more than one file,
+        # or if configured to always total
+        if self.total.n_files > 1 or self.config.always_total:
             self.writeout(rule)
             args = ("TOTAL", self.total.n_statements, self.total.n_missing)
             if self.branches: